+++ CHANGELOG.md
@@ -16,7 +16,7 @@ labeled as 2.7.1. Subsequent releases will follow
   *
   
 ### Fixed
-  *
+  * Fixed abandon command
   *
 
 ## [2.7.22] - 2017-05-11